+++ lib/src/network/socket/mod.rs
@@ -1,3 +1,9 @@
+//! Socket handling is platform specific code.
+//!
+//! The submodules of this module represent the different implementations for
+//! each supported platform.
+//! Depending on the target, the respective platform is read and loaded into this scope.
+
 /// Shared unix stuff
 #[cfg(not(target_os = "windows"))]
 mod unix;