+++ tests/client/follow.rs
@@ -1,7 +1,6 @@
 use std::collections::HashMap;
 
-use anyhow::Context;
-use anyhow::Result;
+use anyhow::{Context, Result};
 
 use crate::fixtures::*;
 use crate::helper::*;