+++ tests/client/mod.rs
@@ -1,4 +1,5 @@
 mod completions;
+mod edit;
 mod follow;
 mod group;
 mod log;