+++ tests/helper/task.rs
@@ -20,22 +20,29 @@ pub async fn add_task(shared: &Shared, command: &str, start_immediately: bool) -
         .context("Failed to to add task.")
 }
 
-/// Adds a task to a specific group of the test daemon.
-pub async fn add_task_to_group(shared: &Shared, command: &str, group: &str) -> Result<Message> {
-    let message = Message::Add(AddMessage {
+/// Create a bare AddMessage for testing.
+/// This is just here to minimize boilerplate code.
+pub fn create_add_message(shared: &Shared, command: &str) -> AddMessage {
+    AddMessage {
         command: command.into(),
         path: shared.pueue_directory(),
         envs: HashMap::new(),
         start_immediately: false,
         stashed: false,
-        group: group.to_owned(),
+        group: PUEUE_DEFAULT_GROUP.to_string(),
         enqueue_at: None,
         dependencies: Vec::new(),
         label: None,
         print_task_id: false,
-    });
+    }
+}
 
-    send_message(shared, message)
+/// Adds a task to a specific group of the test daemon.
+pub async fn add_task_to_group(shared: &Shared, command: &str, group: &str) -> Result<Message> {
+    let mut message = create_add_message(shared, command);
+    message.group = group.to_string();
+
+    send_message(shared, Message::Add(message))
         .await
         .context("Failed to to add task to group.")
 }