+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Interpret the `$EDITOR` command, when editing a task's command/path, as a shell expression instead of an executable ([#336](https://github.com/Nukesor/pueue/issues/336)).
     This gives users more control over how their editor should be started.
 
+### Misc
+
+- Continuation of testing the `pueue` client, pushing the test coverage from ~70% to ~73%.
+
 ## [2.1.0] - 2022-07-21
 
 ### Added